Time: 28 ms (11.21%), Space: 26.9 MB (59.66%) - LeetHub

+++ 1137-n-th-tribonacci-number/1137-n-th-tribonacci-number.cs
@@ -1,16 +1,16 @@
 public class Solution {
     public int Tribonacci(int n) {
-        long a=0;
-        long b=0;
-        long c=1;
+        int a=0;
+        int b=0;
+        int c=1;
         if(n==0) return 0;
         for(int i=1;i<n;i++)
         {
-            long t=c;            
+            int t=c;            
             c=a+b+c;
             a=b;
             b=t;
         }
-        return (int)c;
+        return c;
     }
 }
